dump: block signals with help of ptrace
Currently crtools blocks signals from a parasite code, but for executing
the parasite code a process has to handle all pending non-blockedsignals.

It is not a problem until crtools decide to support stopped tasks.

Registers were saved before blocking signals, so a task can enter in a
signal hangler and block a part of signals, than it will be restored out
of signal handler with blocked signals.