parasite: move code about parasite initialization in a separate function
The current scheme is used two commands to daemonize parasite.
The first command blocks signals, then a sigframe generated and the
second command daemonizes parasite. If CRIU crashes between this
commands, signals will not be unlocked.

PTRACE_GETSIGMASK allows to do the same with one command and a window,
when a process state can't be restored, is smaller.