tcp: add support of intermediate closing states
For restoring this state we need to queue a fin packet in one of both
direction. If we close output direction, a socket turns into FIN-WAIT.
If we close ingoing direction, a socket turns into CLOSE_WAIT.
If we close both directions, a socket turns into onf of LAST_ACK or
CLOSING, it depends on the order of closing.

The size of fin packet is 1, but it can be received, but fin packet
doesn't contain user data. For this reason we need to adjust the size
data in a send queue.

A conntroll message is used for sending fin packet, but it can't contain
real data because of
* fin packet can be sent after real data. In this case we can't return
  an error, because we must return the size of real data, which were sent.
  We can say that fin packet has size 1...
* A send queue is restored in two stages and the repair mode is disabled
  between them, but fin packet can be queued only if a socket is in
  repair mode.

Known issues: sk->shutdown isn't restore yet