criu: correctly handle mixed-permission mapped files
An mmaped file is opened O_RDONLY or O_RDWR depending on the permissions
on the first vma dump_task_mm() encounters mapping that file. This means
that if a file has multiple MAP_SHARED mappings, some of which are
read-only and some of which are read-write, and the first encountered
mapping happens to be read-only, the file will be opened with O_RDONLY
during restore, and mmap(PROT_WRITE) will fail with EACCES.

To fix this, defer filemap open flags selection until restore, since
mapped files are opened per mapping.

Signed-off-by: Jamie Liu <jamieliu@google.com>