restore: look up root for each file
mntns is looked up by mnt_id, then /proc/NS_PID/root is opened.

This part will be optimized in the "mount: add a fast path for
mntns_collect_root" patch and /proc/NS_PID/root will be opened,
only if NS_PID is not the same as in the previous case.